k8s/telemetry/oob-collectors/bmc-sel-collector/code/pip-requirements.txt
confluent-kafka==1.5.0
prometheus-client==0.8.0
PyYAML==5.3.1
requests==2.24.0
urllib3==1.25.10
